Remove VM xml file instead of erroring out.
If the xml file already exists, its best to try and delete it rather
than asking the user to do so. Its safe to do so since the default
location for the xml is in the /tmp directory.
This is helpful for using the script for automation and minimizes human
intervention.

Signed-off-by: Archana Shinde <archana.m.shinde@intel.com>